requirements.txt
# requirements.txt
mysql-connector-python==8.4.0
